+++ src/typantic/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import version
 
-from typantic._decorator import pydantic_to_typer
+from typantic._decorator import add_command, pydantic_to_typer
 
 __version__ = version("typantic")
-__all__ = ["pydantic_to_typer"]
+__all__ = ["add_command", "pydantic_to_typer"]